+++ compiler/ml/typedecl.ml
@@ -49,6 +49,7 @@ type error =
   | Unbound_type_var_ext of type_expr * extension_constructor
   | Varying_anonymous
   | Val_in_structure
+  | Invalid_attribute of string
   | Bad_immediate_attribute
   | Bad_unboxed_attribute of string
   | Boxed_and_unboxed
@@ -288,6 +289,11 @@ let make_constructor env type_path type_params sargs sret_type =
     widen z;
     (targs, Some tret_type, args, Some ret_type, params)
 
+let is_not_undefined_attr (attr : attribute) =
+  match attr with
+  | {Location.txt = "notUndefined"; _}, _ -> true
+  | _ -> false
+
 (* Check that all the variables found in [ty] are in [univ].
    Because [ty] is the argument to an abstract type, the representation
    of that abstract type could be any subexpression of [ty], in particular
@@ -295,6 +301,20 @@ let make_constructor env type_path type_params sargs sret_type =
 *)
 
 let transl_declaration ~type_record_as_object ~untagged_wfc env sdecl id =
+  (* Check for @notUndefined attribute *)
+  let has_not_undefined =
+    List.exists is_not_undefined_attr sdecl.ptype_attributes
+  in
+  (if has_not_undefined then
+     match (sdecl.ptype_kind, sdecl.ptype_manifest) with
+     | Ptype_abstract, None -> ()
+     | _ ->
+       raise
+         (Error
+            ( sdecl.ptype_loc,
+              Invalid_attribute
+                "@notUndefined can only be used on abstract types" )));
+
   (* Bind type parameters *)
   reset_type_variables ();
   Ctype.begin_def ();
@@ -2090,6 +2110,7 @@ let report_error ppf = function
       "The field @{<info>%s@} is defined several times in this record. Fields \
        can only be added once to a record."
       s
+  | Invalid_attribute msg -> fprintf ppf "%s" msg
   | Duplicate_label (s, Some record_name) ->
     fprintf ppf
       "The field @{<info>%s@} is defined several times in the record \