+++ compiler/ml/typedecl.mli
@@ -62,6 +62,7 @@ val is_fixed_type : Parsetree.type_declaration -> bool
 
 (* for typeopt.ml *)
 val get_unboxed_type_representation : Env.t -> type_expr -> type_expr option
+val is_not_undefined_attr : Parsetree.attribute -> bool
 
 type native_repr_kind = Unboxed | Untagged